ZXGraph plotting is now in its own file
