+++ src/tqec/sketchup/zx_graph.py
@@ -4,14 +4,11 @@ from __future__ import annotations
 
 import itertools
 from copy import copy
-from dataclasses import astuple, dataclass
+from dataclasses import dataclass
 from enum import Enum
 from typing import TYPE_CHECKING, cast
 
 import networkx as nx
-import numpy as np
-import numpy.typing as npt
-from mpl_toolkits.mplot3d.axes3d import Axes3D
 
 from tqec.exceptions import TQECException
 from tqec.position import Direction3D, Position3D
@@ -75,11 +72,6 @@ class ZXEdge:
 
 _NODE_DATA_KEY = "tqec_zx_node_data"
 _EDGE_DATA_KEY = "tqec_zx_edge_data"
-
-CORRELATION_COLOR: dict[NodeType, str] = {
-    NodeType.X: "#ff0000",
-    NodeType.Z: "#0000ff",
-}
 
 
 class ZXGraph:
@@ -211,153 +203,6 @@ class ZXGraph:
             data[_EDGE_DATA_KEY]
             for _, _, data in self._graph.edges(position, data=True)
         ]
-
-    def draw_as_zx_graph_on(
-        self,
-        ax: Axes3D,
-        *,
-        node_size: int = 400,
-        hadamard_size: int = 200,
-        edge_width: int = 1,
-    ) -> None:
-        """Draw the 3D graph using matplotlib on the provided figure.
-
-        Args:
-            ax: a 3-dimensional ax to draw on.
-            node_size: The size of the node. Default is 400.
-            hadamard_size: The size of the Hadamard transition. Default is 200.
-            edge_width: The width of the edge. Default is 1.
-        """
-        non_virtual_nodes = [n for n in self.nodes if not n.is_virtual]
-        non_virtual_nodes_array = _positions_array(
-            *[n.position for n in non_virtual_nodes]
-        )
-        if non_virtual_nodes_array.size > 0:
-            ax.scatter(
-                *non_virtual_nodes_array,
-                s=node_size,
-                c=[
-                    "black" if n.node_type == NodeType.X else "white"
-                    for n in non_virtual_nodes
-                ],
-                alpha=1.0,
-                edgecolors="black",
-            )
-
-        for edge in self.edges:
-            pos_array = _positions_array(edge.u.position, edge.v.position)
-            ax.plot(
-                *pos_array,
-                color="tab:gray",
-                linewidth=edge_width,
-            )
-            if edge.has_hadamard:
-                hadamard_position = np.mean(pos_array, axis=1)
-                # use yellow square to indicate Hadamard transition
-                ax.scatter(
-                    *hadamard_position,
-                    s=hadamard_size,
-                    c="yellow",
-                    alpha=1.0,
-                    edgecolors="black",
-                    marker="s",
-                )
-        ax.grid(False)
-        for dim in (ax.xaxis, ax.yaxis, ax.zaxis):
-            dim.set_ticks([])
-        x_limits, y_limits, z_limits = ax.get_xlim3d(), ax.get_ylim3d(), ax.get_zlim3d()  # type: ignore
-
-        plot_radius = 0.5 * max(
-            abs(limits[1] - limits[0]) for limits in [x_limits, y_limits, z_limits]
-        )
-
-        ax.set_xlim3d(
-            [np.mean(x_limits) - plot_radius, np.mean(x_limits) + plot_radius]
-        )
-        ax.set_ylim3d(
-            [np.mean(y_limits) - plot_radius, np.mean(y_limits) + plot_radius]
-        )
-        ax.set_zlim3d(
-            [np.mean(z_limits) - plot_radius, np.mean(z_limits) + plot_radius]
-        )
-
-    def draw_as_correlation_surface_on(
-        self,
-        ax: Axes3D,
-        correlation_edge_width: int = 3,
-    ) -> None:
-        for edge in self.edges:
-            pos_array = _positions_array(edge.u.position, edge.v.position)
-            if not edge.has_hadamard:
-                correlation_type = edge.u.node_type
-                ax.plot(
-                    *pos_array,
-                    color=CORRELATION_COLOR[correlation_type],
-                    linewidth=correlation_edge_width,
-                )
-            else:
-                hadamard_position = np.mean(pos_array, axis=1)
-                for node in [edge.u, edge.v]:
-                    ax.plot(
-                        *np.hstack(
-                            [
-                                hadamard_position.reshape(3, 1),
-                                _positions_array(node.position),
-                            ]
-                        ),
-                        color=CORRELATION_COLOR[node.node_type],
-                        linewidth=correlation_edge_width,
-                    )
-
-    def draw(
-        self,
-        *,
-        show_correlation_subgraph_index: int | None = None,
-        figsize: tuple[float, float] = (5, 6),
-        title: str | None = None,
-        node_size: int = 400,
-        hadamard_size: int = 200,
-        edge_width: int = 1,
-        correlation_edge_width: int = 3,
-    ) -> None:
-        """Draw the 3D graph using matplotlib.
-
-        Args:
-            show_correlation_subgraph_index: The index of the correlation subgraph to show.
-            figsize: The figure size. Default is (5, 6).
-            title: The title of the plot. Default is the name of the graph.
-            node_size: The size of the node. Default is 400.
-            hadamard_size: The size of the Hadamard transition. Default is 200.
-            edge_width: The width of the edge. Default is 1.
-            correlation_edge_width: The width of the correlation edge. Default is 3.
-        """
-        import matplotlib.pyplot as plt
-
-        fig = plt.figure(figsize=figsize)
-        ax = cast(Axes3D, fig.add_subplot(111, projection="3d"))
-
-        self.draw_as_zx_graph_on(
-            ax, node_size=node_size, hadamard_size=hadamard_size, edge_width=edge_width
-        )
-
-        if show_correlation_subgraph_index is not None:
-            correlation_subgraphs = self.find_correlation_subgraphs()
-            if show_correlation_subgraph_index >= len(correlation_subgraphs):
-                raise TQECException(
-                    f"Only {len(correlation_subgraphs)} correlation subgraphs found."
-                    f"Index {show_correlation_subgraph_index} is out of range."
-                )
-            correlation_subgraph = correlation_subgraphs[
-                show_correlation_subgraph_index
-            ]
-
-            correlation_subgraph.draw_as_correlation_surface_on(
-                ax, correlation_edge_width
-            )
-
-        ax.set_title(title or self.name)
-        fig.tight_layout()
-        plt.show()
 
     def to_block_graph(self, name: str = "") -> BlockGraph:
         """Construct a block graph from a ZX graph.
@@ -518,7 +363,3 @@ class ZXGraph:
         return single_node_correlation_subgraphs + list(
             multi_edges_correlation_subgraphs.values()
         )
-
-
-def _positions_array(*positions: Position3D) -> npt.NDArray[np.int_]:
-    return np.array([astuple(p) for p in positions]).T
